feat: Updated individual component for input an text
